+++ changelog.md
@@ -3,12 +3,23 @@
 ---
 Also see: [Architect changelog](https://github.com/architect/architect/blob/master/changelog.md)
 ---
+
+## [On branch]
+
+### Added
+
+- Added fallback event object for direct invocation of event functions using `arc.events.subscribe`
+
+---
+
 ## [3.5.0] 2020-01-24
 
 ### Added
 
 - `arc.queues.publish` can now accept `groupID` for fifo queues
 - erase session with `return {session: {}, location: '/'}` syntax
+
+---
 
 ## [3.4.14] 2019-12-13